+++ src/main/java/com/doublew2w/sbs/mybatis/executor/statement/SimpleStatementHandler.java
@@ -67,8 +67,8 @@ public class SimpleStatementHandler extends BaseStatementHandler {
 
   @Override
   public <E> List<E> query(Statement statement, ResultHandler resultHandler) throws SQLException {
-    log.info("执行查询");
     String sql = boundSql.getSql();
+    log.info("sql:{}", sql);
     statement.execute(sql);
     return resultSetHandler.handleResultSets(statement);
   }