+++ src/test/java/com/doublew2w/sbs/mybatis/test/IActivityDaoApiTest.java
@@ -8,7 +8,11 @@ import com.doublew2w.sbs.mybatis.session.SqlSessionFactoryBuilder;
 import com.doublew2w.sbs.mybatis.test.dao.IActivityDao;
 import com.doublew2w.sbs.mybatis.test.po.Activity;
 import java.io.IOException;
+import java.lang.reflect.Member;
+import java.lang.reflect.Modifier;
+import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
+import ognl.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -55,5 +59,46 @@ public class IActivityDaoApiTest {
     sqlSession.commit();
 
     log.info("测试结果：count：{} idx：{}", res, JSON.toJSONString(activity.getId()));
+  }
+
+  @Test
+  public void test_branch15_query() {
+    // 2. 获取映射器对象
+    IActivityDao dao = sqlSession.getMapper(IActivityDao.class);
+    // 3. 测试验证
+    Activity req = new Activity();
+    req.setActivityId(100001L);
+    Activity res = dao.queryActivityByIdForDynamicSql(req);
+    log.info("测试结果：{}", JSON.toJSONString(res));
+  }
+
+  @Test
+  public void test_ognl() throws OgnlException {
+    Activity req = new Activity();
+    req.setActivityId(1L);
+    req.setActivityName("测试活动");
+    req.setActivityDesc("小傅哥的测试内容");
+
+    OgnlContext context =
+        new OgnlContext(
+            new AbstractMemberAccess() {
+              @Override
+              public boolean isAccessible(
+                  Map context, Object target, Member member, String propertyName) {
+                int modifiers = member.getModifiers();
+                return Modifier.isPublic(modifiers);
+              }
+            },
+            null,
+            null,
+            null);
+    context.setRoot(req);
+    Object root = context.getRoot();
+
+    Object activityName = Ognl.getValue("activityName", context, root);
+    Object activityDesc = Ognl.getValue("activityDesc", context, root);
+    Object value = Ognl.getValue("activityDesc.length()", context, root);
+
+    System.out.println(activityName + "\t" + activityDesc + " length：" + value);
   }
 }