+++ src/test/java/com/doublew2w/sbs/mybatis/test/dao/IActivityDao.java
@@ -11,4 +11,6 @@ public interface IActivityDao {
   Activity queryActivityById(Long activityId);
 
   Integer insert(Activity activity);
+
+  Activity queryActivityByIdForDynamicSql(Activity Activity);
 }